Move main GUI entry point to repo root
