+++ gui/drawing_helper.py
