+++ sysml/sysml_spec.py
