+++ tests/test_actions.py
@@ -2,7 +2,7 @@ import unittest
 import sys
 import os
 sys.path.append(os.path.dirname(os.path.dirname(__file__)))
-from sysml_repository import SysMLRepository
+from sysml.sysml_repository import SysMLRepository
 
 class ActionNameTests(unittest.TestCase):
     def setUp(self):
